echoes_mvp/client/requirements.txt

opencv-python
mediapipe==0.10.21
speechrecognition
pyaudio
python-socketio
